Divide tests into multiple files
